Update README.md
Predicting and finding proactive ways of reducing attrition using technologies of your choice For eg: use available data/ and create modeling tools to predict which employees are more likely to leave given some attributes
The objective of the present report to study factor like businesstravel, department, distance from home, policies and procedures, recognition, appreciation, suggestions of the employee’s by which it helps to know the attrition level in the organizations and factors relating to retain them. This study also helps to find out where the organizations are lagging in retaining.